Changed ome quad_ker, related imports in src and tests
