+++ src/eko/evolution_operator/quad_ker.py
@@ -1,5 +1,6 @@
 r"""Integration kernels."""
 
+import enum
 import logging
 
 import numba as nb
@@ -8,6 +9,9 @@ import numpy as np
 import ekore.anomalous_dimensions.polarized.space_like as ad_ps
 import ekore.anomalous_dimensions.unpolarized.space_like as ad_us
 import ekore.anomalous_dimensions.unpolarized.time_like as ad_ut
+import ekore.operator_matrix_elements.polarized.space_like as ome_ps
+import ekore.operator_matrix_elements.unpolarized.space_like as ome_us
+import ekore.operator_matrix_elements.unpolarized.time_like as ome_ut
 
 from .. import interpolation, mellin
 from .. import scale_variations as sv
@@ -19,6 +23,7 @@ from ..kernels import valence_qed as qed_v
 from ..matchings import lepton_number
 from ..scale_variations import expanded as sv_expanded
 from ..scale_variations import exponentiated as sv_exponentiated
+from ..scale_variations.exponentiated import gamma_variation
 
 logger = logging.getLogger(__name__)
 
@@ -104,6 +109,67 @@ def select_QEDvalence_element(ker, mode0, mode1):
     return ker[index1, index2]
 
 
+class MatchingMethods(enum.IntEnum):
+    """Enumerate matching methods."""
+
+    FORWARD = enum.auto()
+    BACKWARD_EXACT = enum.auto()
+    BACKWARD_EXPANDED = enum.auto()
+
+
+@nb.njit(cache=True)
+def build_ome(A, matching_order, a_s, backward_method):
+    r"""Construct the matching expansion in :math:`a_s` with the appropriate
+    method.
+
+    Parameters
+    ----------
+    A : numpy.ndarray
+        list of |OME|
+    matching_order : tuple(int,int)
+        perturbation matching order
+    a_s : float
+        strong coupling, needed only for the exact inverse
+    backward_method : MatchingMethods
+        empty or method for inverting the matching condition (exact or expanded)
+
+    Returns
+    -------
+    ome : numpy.ndarray
+        matching operator matrix
+    """
+    # to get the inverse one can use this FORM snippet
+    # Symbol a;
+    # NTensor c,d,e;
+    # Local x=-(a*c+a**2* d + a**3 * e);
+    # Local bi = 1+x+x**2+x**3;
+    # Print;
+    # .end
+    ome = np.eye(len(A[0]), dtype=np.complex128)
+    A = A[:, :, :]
+    A = np.ascontiguousarray(A)
+    if backward_method is MatchingMethods.BACKWARD_EXPANDED:
+        # expended inverse
+        if matching_order[0] >= 1:
+            ome -= a_s * A[0]
+        if matching_order[0] >= 2:
+            ome += a_s**2 * (-A[1] + A[0] @ A[0])
+        if matching_order[0] >= 3:
+            ome += a_s**3 * (-A[2] + A[0] @ A[1] + A[1] @ A[0] - A[0] @ A[0] @ A[0])
+    else:
+        # forward or exact inverse
+        if matching_order[0] >= 1:
+            ome += a_s * A[0]
+        if matching_order[0] >= 2:
+            ome += a_s**2 * A[1]
+        if matching_order[0] >= 3:
+            ome += a_s**3 * A[2]
+        # need inverse exact ?  so add the missing pieces
+        if backward_method is MatchingMethods.BACKWARD_EXACT:
+            ome = np.linalg.inv(ome)
+    return ome
+
+
 spec = [
     ("is_singlet", nb.boolean),
     ("is_QEDsinglet", nb.boolean),
@@ -176,7 +242,7 @@ class QuadKerBase:
 
 
 @nb.njit(cache=True)
-def quad_ker(
+def quad_ker_ad(
     u,
     order,
     mode0,
@@ -588,3 +654,102 @@ def quad_ker_qed(
                 * ker
             )
     return ker
+
+
+@nb.njit(cache=True)
+def quad_ker_ome(
+    u,
+    order,
+    mode0,
+    mode1,
+    is_log,
+    logx,
+    areas,
+    a_s,
+    nf,
+    L,
+    sv_mode,
+    Lsv,
+    backward_method,
+    is_msbar,
+    is_polarized,
+    is_time_like,
+):
+    r"""Raw kernel inside quad.
+
+    Parameters
+    ----------
+    u : float
+        quad argument
+    order : tuple(int,int)
+        perturbation matching order
+    mode0 : int
+        pid for first element in the singlet sector
+    mode1 : int
+        pid for second element in the singlet sector
+    is_log : boolean
+        logarithmic interpolation
+    logx : float
+        Mellin inversion point
+    areas : tuple
+        basis function configuration
+    a_s : float
+        strong coupling, needed only for the exact inverse
+    nf: int
+        number of active flavor below threshold
+    L : float
+        :math:``\ln(\mu_F^2 / m_h^2)``
+    backward_method : InversionMethod or None
+        empty or method for inverting the matching condition (exact or expanded)
+    is_msbar: bool
+        add the |MSbar| contribution
+    is_polarized : boolean
+        is polarized evolution ?
+    is_time_like : boolean
+        is time-like evolution ?
+
+    Returns
+    -------
+    ker : float
+        evaluated integration kernel
+    """
+    ker_base = QuadKerBase(u, is_log, logx, mode0)
+    integrand = ker_base.integrand(areas)
+    if integrand == 0.0:
+        return 0.0
+    # compute the ome
+    if ker_base.is_singlet or ker_base.is_QEDsinglet:
+        indices = {21: 0, 100: 1, 90: 2}
+        if is_polarized:
+            if is_time_like:
+                raise NotImplementedError("Polarized, time-like is not implemented")
+            A = ome_ps.A_singlet(order, ker_base.n, nf, L)
+        else:
+            if is_time_like:
+                A = ome_ut.A_singlet(order, ker_base.n, L)
+            else:
+                A = ome_us.A_singlet(order, ker_base.n, nf, L, is_msbar)
+    else:
+        indices = {200: 0, 91: 1}
+        if is_polarized:
+            if is_time_like:
+                raise NotImplementedError("Polarized, time-like is not implemented")
+            A = ome_ps.A_non_singlet(order, ker_base.n, L)
+        else:
+            if is_time_like:
+                A = ome_ut.A_non_singlet(order, ker_base.n, L)
+            else:
+                A = ome_us.A_non_singlet(order, ker_base.n, nf, L)
+
+    # correct for scale variations
+    if sv_mode == sv.Modes.exponentiated:
+        A = gamma_variation(A, order, nf, Lsv)
+
+    # build the expansion in alpha_s depending on the strategy
+    ker = build_ome(A, order, a_s, backward_method)
+
+    # select the needed matrix element
+    ker = ker[indices[mode0], indices[mode1]]
+
+    # recombine everything
+    return np.real(ker * integrand)
